pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "cofone"
version = "0.1.0"
description = "Simple, fast RAG library for Python"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Il tuo nome", email = "tua@email.com"}]
dependencies = ["openai>=1.0.0", "python-dotenv>=1.0.0", "httpx>=0.27.0", "beautifulsoup4>=4.12.0", "pydantic"]

[project.optional-dependencies]
pdf   = ["pypdf>=4.0.0"]
faiss = ["faiss-cpu>=1.7.0", "sentence-transformers>=2.0.0"]
web   = ["wikipedia>=1.4.0", "youtube-transcript-api>=0.6.0"]
all   = ["pypdf>=4.0.0", "faiss-cpu>=1.7.0", "sentence-transformers>=2.0.0", "wikipedia>=1.4.0", "youtube-transcript-api>=0.6.0"]